src/errors/exception_handlers.py
from fastapi import FastAPI
from fastapi.exceptions import RequestValidationError
from starlette import status
from starlette.responses import JSONResponse

from src.errors.errors import ApiError


def setup_exception_handlers(app: FastAPI) -> None:
    @app.exception_handler(ApiError)
    async def api_error_handler(_, error: ApiError) -> JSONResponse:
        return JSONResponse(
            status_code=error.status_code,
            content={
                "message": error.message,
                "version": app.version
            }
        )

    @app.exception_handler(RequestValidationError)
    async def api_error_handler(_, error: RequestValidationError) -> JSONResponse:
        return JSONResponse(
            status_code=status.HTTP_400_BAD_REQUEST,
            content=error.errors()
        )
