requirements.txt
fastapi
python-dotenv==1.0.1
SQLAlchemy==2.0.29
uvicorn==0.29.0
pytest
pytest-cov
httpx
pydantic~=2.9.2
pydantic[email]
psycopg2-binary
testcontainers[postgresql]
pystache
requests